src/TriggerTextract/requirements.txt
boto3
textract-trp